Pull in latest core versions and add migrations to/from both
