+++ startos/install/versions/index.ts
@@ -1,4 +1,5 @@
-export { v29_2_0_9 as knotsCurrent } from './v29_2_0_9-beta.0'
-import { v29_1_0_B1 } from './v29_1_0_B1'
+export { v29_2_0_9 as knotsCurrent } from './v29_2_0_9-beta.1'
+import { v29_2_0_2 as core29_2 } from 'bitcoin-core-startos/startos/install/versions/v29.2.0_2-beta.4'
+import { v30_0_0_1 as coreCurrent } from 'bitcoin-core-startos/startos/install/versions/v30.0.0_1-beta.7'
 
-export const other = [v29_1_0_B1]
+export const other = [coreCurrent, core29_2]
